Create BITMAP - Bitmap
I have used BFS to solve this problem. The idea is to push all 1's in the input in a queue. Then start using BFS and look in the up, down, left and right direction of the current node. If the new position has a distance of greater than current_distance+1 then push it in the queue and update its value to current_distance+1. We initially made the distance of all the 1's in the array as 0 (since the node is 1 itself thus its distance from nearest 1 is zero i.e. its own location) and INT_MAX for all 0's(INT_MAX is the maximum value that an integer can hold. We can use a very large number also but this method seems more convinient).